pyproject.toml
[project]
name = "vehicle-homologation-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.0",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.1.0",
    "langchain-huggingface>=1.1.0",
    "langchain-openai>=1.1.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.0",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
]

[tool.pytest.ini_options]
pythonpath = "."
